Split discount fee calculator from fee calculator
