+++ src/main/java/nextstep/subway/domain/service/StationPathDiscountFeeContext.java
@@ -1,0 +1,11 @@
+package nextstep.subway.domain.service;
+
+import lombok.Builder;
+import lombok.Getter;
+import nextstep.member.domain.Member;
+
+@Getter
+@Builder
+public class StationPathDiscountFeeContext {
+    private Member member;
+}
