+++ src/main/java/nextstep/subway/domain/service/StationPathFeeContext.java
@@ -2,7 +2,6 @@ package nextstep.subway.domain.service;
 
 import lombok.Builder;
 import lombok.Getter;
-import nextstep.member.domain.Member;
 import nextstep.subway.domain.StationLine;
 
 import java.math.BigDecimal;
@@ -12,6 +11,5 @@ import java.util.List;
 @Builder
 public class StationPathFeeContext {
     private BigDecimal distance;
-    private Member member;
     private List<StationLine> lines;
 }
