+++ src/test/java/nextstep/subway/unit/StationDiscountFeeCalculatorServiceTest.java
@@ -1,0 +1,57 @@
+package nextstep.subway.unit;
+
+import nextstep.member.domain.Member;
+import nextstep.member.fixture.MemberSpec;
+import nextstep.subway.domain.service.*;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import java.math.BigDecimal;
+import java.math.RoundingMode;
+
+public class StationDiscountFeeCalculatorServiceTest {
+    private UserAgeDiscountFeeCalculator userAgeDiscountFeeCalculator = new UserAgeDiscountFeeCalculator();
+
+    @DisplayName("청소년 할인 적용된 요금 계산")
+    @Test
+    void calculateFeeTest_Teenager_DiscountFee() {
+        //given
+        final BigDecimal totalFee = BigDecimal.valueOf(1250);
+        final Member member = MemberSpec.of("TEST@nextstep.com", 18);
+
+        final StationPathDiscountFeeContext context = StationPathDiscountFeeContext.builder()
+                .member(member)
+                .build();
+
+        //when
+        final BigDecimal discountFee = userAgeDiscountFeeCalculator.calculateDiscountFee(totalFee, context);
+
+        //then
+        final BigDecimal twentyPercent = BigDecimal.valueOf(20).divide(BigDecimal.valueOf(100), 1, RoundingMode.DOWN);
+        final BigDecimal expectedDiscountFee = totalFee.subtract(BigDecimal.valueOf(250)).multiply(twentyPercent);
+
+        Assertions.assertEquals(0, discountFee.compareTo(expectedDiscountFee));
+    }
+
+    @DisplayName("어린이 할인 적용된 요금 계산")
+    @Test
+    void calculateFeeTest_Children_DiscountFee() {
+        //given
+        final BigDecimal totalFee = BigDecimal.valueOf(1250);
+        final Member member = MemberSpec.of("TEST@nextstep.com", 6);
+
+        final StationPathDiscountFeeContext context = StationPathDiscountFeeContext.builder()
+                .member(member)
+                .build();
+
+        //when
+        final BigDecimal discountFee = userAgeDiscountFeeCalculator.calculateDiscountFee(totalFee, context);
+
+        //then
+        final BigDecimal fiftyPercent = BigDecimal.valueOf(50).divide(BigDecimal.valueOf(100), 1, RoundingMode.DOWN);
+        final BigDecimal expectedFee = totalFee.subtract(BigDecimal.valueOf(250)).multiply(fiftyPercent);
+
+        Assertions.assertEquals(0, discountFee.compareTo(expectedFee));
+    }
+}
